+++ CHANGES.rst
@@ -4,6 +4,7 @@ Version 2.3.0
 Unreleased
 
 -   Drop support for Python 3.8 and 3.9.
+-   Remove previously deprecated code.
 
 
 Version 2.2.0